Added function to check denoising vs number of components in the mixture + some experiments
All changes are in the folder /fourth assignment/

1. a new function - check_psnr_vs_num_of_components.m receives a picture
name and denoises it using 1-10 component mixtures. it returns a vector
of the psnr we got in each denoising.
2. experinments were done in /fourth assignment/eperiments. In this
folder you can find experiments of PSNR vs. num of components of the
peppers image and the tiger image.
3. another folder in the experiments folder holds the visualization of
the learnt covariance matrices for 1,5,10 component mixture we learnt.
4. All this was summerized in the Lyx file in fourth assignments folder
called Num_of_components.lyx. Note that this file is not final.